// apps/web/src/lib/generated/api/ChatMessage.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ToolCall } from './ToolCall';

/**
 * Chat message (user or assistant).
 */
export type ChatMessage = {
  id: string;
  conversation_id: string;
  /**
   * "user" | "assistant" | "system" | "tool"
   */
  role: string;
  content: string;
  tool_calls: Array<ToolCall> | null;
  created_at: string;
};